revert a493818. remove Dashboard and related components from safe-framework
