+++ examples/js/fixtures.js
@@ -623,115 +623,3 @@ export const scatterSeries = {
     data: maleData
   }]
 }
-
-// //////////// Dashboard Data ////////////////
-
-export const dashboardData = {
-  _id: '1',
-  title: 'Visualizations',
-  visualizations: [{
-    _id: 'columnViz',
-    name: 'TrafficViolations By Gender',
-    source: {
-      _id: '1',
-      name: 'TrafficViolationsData',
-      fields: []
-    },
-    visualizationType: {
-      _id: '1',
-      name: 'Column'
-    },
-    visualizationParams: {
-      title: {
-        text: 'Counts by Gender'
-      },
-      series: 'series',
-      xAxis: 'name',
-      yAxis: 'y'
-    },
-    filters: []
-  }, {
-    _id: 'mapViz',
-    name: 'Sample Map: Map View',
-    source: {
-      _id: '2',
-      name: 'CSV_20160122',
-      fields: [{
-        name: 'Age',
-        dataType: 'Number'
-      }, {
-        name: 'County',
-        dataType: 'String'
-      }, {
-        name: 'Height',
-        dataType: 'Number'
-      }, {
-        name: 'Latitude',
-        dataType: 'Number'
-      }, {
-        name: 'Longitude',
-        dataType: 'Number'
-      }, {
-        name: 'Street Address',
-        dataType: 'String'
-      }, {
-        name: 'Weight',
-        dataType: 'Number'
-      }]
-    },
-    visualizationType: {
-      _id: '2',
-      name: 'Map'
-    },
-    visualizationParams: {
-      title: 'Simple Map',
-      destinationPrefix: 'to',
-      latField: 'latitude',
-      longField: 'longitude',
-      label: [
-        'label'
-      ],
-      sourcePrefix: 'from'
-    },
-    filters: []
-  }, {
-    _id: 'tableViz',
-    name: 'Sample Map: Table View',
-    source: {
-      _id: '2',
-      name: 'CSV_20160122',
-      fields: [{
-        name: 'Age',
-        dataType: 'Number'
-      }, {
-        name: 'County',
-        dataType: 'String'
-      }, {
-        name: 'Height',
-        dataType: 'Number'
-      }, {
-        name: 'Latitude',
-        dataType: 'Number'
-      }, {
-        name: 'Longitude',
-        dataType: 'Number'
-      }, {
-        name: 'Street Address',
-        dataType: 'String'
-      }, {
-        name: 'Weight',
-        dataType: 'Number'
-      }]
-    },
-    visualizationType: {
-      _id: '3',
-      name: 'Table'
-    },
-    visualizationParams: {
-      title: {
-        text: 'Map'
-      }
-    },
-    filters: []
-  }]
-}
