--- src/Dashboard/index.js
@@ -1,8 +1,0 @@
-import {Dashboard} from './Dashboard'
-import {Visualization} from './Visualization'
-import {VisualizationToolbar} from './VisualizationToolbar'
-
-export default Dashboard
-
-export {Visualization}
-export {VisualizationToolbar}
