+++ src/index.js
@@ -13,7 +13,6 @@ import SelectField from './SelectField'
 import AreaChart, {DefaultAreaChart} from './AreaChart'
 import BarChart, {DefaultBarChart} from './BarChart'
 import ColumnChart, {DefaultColumnChart} from './ColumnChart'
-import Dashboard, {Visualization} from './Dashboard'
 import LineChart, {DefaultLineChart} from './LineChart'
 import PieChart, {DefaultPieChart} from './PieChart'
 import ScatterPlot, {DefaultScatterPlot} from './ScatterPlot'
@@ -23,7 +22,6 @@ export {AreaChart}
 export {BarChart}
 export {BasicDataTable}
 export {ColumnChart}
-export {Dashboard}
 export {DataTable}
 export {DefaultAreaChart}
 export {DefaultBarChart}
@@ -45,7 +43,6 @@ export {ScatterPlot}
 export {Section}
 export {Sections}
 export {SelectField}
-export {Visualization}
 export {Wrapper}
 
 export default {
@@ -53,7 +50,6 @@ export default {
   BarChart,
   BasicDataTable,
   ColumnChart,
-  Dashboard,
   DataTable,
   DefaultAreaChart,
   DefaultColumnChart,
